sd-gui: pre-align shapes to pixel grid and tweak antialiasing to avoid aliasing artifacts
We give up a bit of 'crispness' in our lines in exchange, but the
blobbing artifacts and jaggedness in the interactions between line
segments and bezier curves (often, coincidentally straight) are gone.

There is a bad interaction between our manual alignment and egui::Scene,
as it is not able to perform pixel-grid alignment after panzoom
transformation, but it seems like this is the best we can do given
egui's current Scene model.

Ideally, we would perform the pixel-grid alignment after the Scene
panzoom transformation is applied, but this probably requires changes in
egui's tessellator - for now, we have told the tessellator to not
automatically perform this alignment on LineSegments (as we want
consistency with CubicBeziers).

Doubling the feathering size also seems to alleviate the 'blobbing'
artifacts where lines connect.